Expose ack_fun through the init_info
`commit_fun` is already exposed, but `ack_fun` is not. This is useful
when acking does not happen inside the worker process.

E.g. We divide messages between multiple processes and later in a
separate process we create a transaction and commit a batch. Now we need
to let the consumer know that it's okay to fetch more messages.